Support repeating_trip CRUD (C.14 - C.15)
