+++ app/controllers/trips_controller.rb
@@ -349,17 +349,6 @@ class TripsController < ApplicationController
       :pickup_address_id,
       :pickup_time,
       :provider_id, # We normally wouldn't accept this and would set it manually on the instance, but in this controller we're setting it in the params dynamically
-      :repeats_sundays,
-      :repeats_mondays,
-      :repeats_tuesdays,
-      :repeats_wednesdays,
-      :repeats_thursdays,
-      :repeats_fridays,
-      :repeats_saturdays,
-      :repetition_customer_informed,
-      :repetition_driver_id,
-      :repetition_interval,
-      :repetition_vehicle_id,
       :run_id,
       :cab,
       :service_level_id,
