+++ app/models/repeating_trip.rb
@@ -2,20 +2,71 @@
 # repeating instance of itself when instructed to. Validation is nonexistent 
 # since all data should already have been vetted by the Trip instance.
 class RepeatingTrip < ActiveRecord::Base
+  include RequiredFieldValidatorModule
   include RecurringRideCoordinatorScheduler
+  include RecurringRideCoordinator
+  include TripCore
 
   has_paper_trail
 
-  belongs_to :customer, -> { with_deleted }
-  belongs_to :driver, -> { with_deleted }
-  belongs_to :dropoff_address, -> { with_deleted }, class_name: "Address"
-  belongs_to :funding_source, -> { with_deleted }
-  belongs_to :mobility, -> { with_deleted }
-  belongs_to :pickup_address, -> { with_deleted }, class_name: "Address"
-  belongs_to :provider, -> { with_deleted }
-  belongs_to :vehicle, -> { with_deleted }
-  belongs_to :trip_purpose, -> { with_deleted }
+  schedules_occurrences_with :repeating_trip,
+    with_attributes: -> (trip) {
+      attrs = {}
+      RepeatingTrip.ride_coordinator_attributes.each {|attr| attrs[attr] = trip.send(attr) }
+      attrs['driver_id'] = trip.repetition_driver_id
+      attrs['vehicle_id'] = trip.repetition_vehicle_id
+      attrs['schedule_attributes'] = {
+        repeat:        1,
+        interval_unit: "week", 
+        start_date:    trip.pickup_time.to_date.to_s,
+        interval:      trip.repetition_interval, 
+        monday:        trip.repeats_mondays    ? 1 : 0,
+        tuesday:       trip.repeats_tuesdays   ? 1 : 0,
+        wednesday:     trip.repeats_wednesdays ? 1 : 0,
+        thursday:      trip.repeats_thursdays  ? 1 : 0,
+        friday:        trip.repeats_fridays    ? 1 : 0,
+        saturday:      trip.repeats_saturdays  ? 1 : 0,
+        sunday:        trip.repeats_sundays    ? 1 : 0
+      }
+      attrs
+    },
+    destroy_future_occurrences_with: -> (trip) {
+      # Be sure not delete occurrences that have already happened.
+      trips = if trip.pickup_time < Time.zone.now
+        Trip.repeating_based_on(trip.repeating_trip).after_today.not_called_back
+      else 
+        Trip.repeating_based_on(trip.repeating_trip).after(trip.pickup_time).not_called_back
+      end
 
+      schedule = trip.repeating_trip.schedule
+      Trip.transaction do
+        trips.find_each do |t|
+          t.destroy unless schedule.occurs_on?(t.pickup_time)
+        end
+      end
+    },
+    destroy_all_future_occurrences_with: -> (trip) {
+      # Be sure not delete occurrences that have already happened.
+      trips = if trip.pickup_time < Time.zone.now
+        Trip.repeating_based_on(trip.repeating_trip).after_today.not_called_back
+      else 
+        Trip.repeating_based_on(trip.repeating_trip).after(trip.pickup_time).not_called_back
+      end
+
+      trips.destroy_all
+    },
+    unlink_past_occurrences_with: -> (trip) {
+      if trip.pickup_time < Time.zone.now
+        Trip.repeating_based_on(trip.repeating_trip).today_and_prior.update_all "repeating_trip_id = NULL"
+      else 
+        Trip.repeating_based_on(trip.repeating_trip).prior_to(trip.pickup_time).update_all "repeating_trip_id = NULL"
+      end
+    }
+
+
+  belongs_to :driver, -> { with_deleted } 
+  belongs_to :vehicle, -> { with_deleted }
+ 
   scope :active, -> { where("(start_date is NULL or start_date <= ?) AND (end_date is NULL or end_date >= ?)", Date.today, Date.today) }
   
   def instantiate!
@@ -48,5 +99,9 @@ class RepeatingTrip < ActiveRecord::Base
     active = false if (start_date && today < start_date) || (end_date && today > end_date)
 
     active
+  end
+
+  def name
+    
   end
 end
