--- app/views/layouts/repeating_trips.html.erb
@@ -1,7 +1,0 @@
-<% content_for :menu_bar do %>
-  <ul id="subnav" class="selfclear">
-    <li><%= link_to translate_helper("view_repeating_trips_link_text"), :action=>:index %></li>
-  </ul>
-<% end %>
-
-<%= render :file => 'layouts/application' %>
