+++ app/views/repeating_trips/_customer.html.erb
@@ -1,0 +1,107 @@
+<%- if @customer.try(:id).present? -%>
+  <div id="trip_customer_name_wrapper">
+    <strong id="trip_customer_name_string"><%= link_to @customer.name, @customer, :target => "_blank" %></strong>
+    <%- if edit_mode != :show %>
+      &nbsp;&nbsp;&nbsp;&nbsp;(<a href="#" id="change_trip_customer_link"><%= translate_helper("change_customer") %></a>)
+    <% end %>
+  </div>
+<% end %>
+<%= autocomplete_field_tag 'customer_name', @customer.try(:name), autocomplete_customers_path(active_only: true), 
+    :size => 30, 
+    :delay => 500, 
+    :update_elements => {
+      :id                        => '#repeating_trip_customer_id,#repeating_trip_customer_attributes_id',
+      :group                        => '#trip_group',
+      :message                   => "#customer_flag_message"
+    },
+    :class => "form-control",
+    :style => @customer.try(:id).present? ? 'display: none' : '', :autocomplete => "off"
+%>
+<%- if edit_mode != :show -%>
+  <%= link_to translate_helper("new"), new_customer_path, :target => "_blank", id: 'new_customer', class: 'btn action-button', style: (@customer.try(:id).present? ? 'display: none;' : '') %>
+  <button type='button' id='view_customer' class='btn action-button' style="<%= (@customer.try(:id).present? ? '' : 'display: none;') %>">
+    <%= translate_helper("view_customer_profile") %>
+  </button>
+<%- end -%>
+<%= javascript_tag do %>
+  $(document).ready(function(){
+    $('#change_trip_customer_link').click(function(event){
+      event.preventDefault();
+      $('#trip_customer_name_wrapper').hide();
+      $('#customer_name').show();
+      $('#new_customer').show();
+    });
+
+    function popupFlagMessage(message) {
+      if(message) {
+        $('#customerFlagMessageText').text(message);
+        $('#customerFlagMessageDialog').modal('show');
+      }
+    }
+
+    function updateUponCustomerChange(customerData) {
+      $('#repeating_trip_customer_attributes_phone_number_1').val(customerData.phone_number_1); 
+      $('#alt_phone').val(customerData.phone_number_2);
+      $('#mobility_notes').val(customerData.mobility_notes);
+      $('#repeating_trip_mobility_id').val(customerData.mobility_id);
+      $('#customer_private_notes').val(customerData.private_notes);
+      $("#repeating_trip_pickup_address_id").val(customerData.address_id);
+      $("#pickup_address").val(customerData.address);
+      $("#repeating_trip_funding_source_id").val(customerData.default_funding_source_id);
+      $("#repeating_trip_service_level").val(customerData.default_service_level);
+      $('#repeating_trip_medicaid_eligible').attr('checked',customerData.medicaid_eligible);
+
+      var elig_data = customerData.customer_eligibilities;
+      if(elig_data && elig_data.length > 0) {
+        $('.customer-eligibilities').html('');
+        $('.customer-eligibilities').append("<h4 class='col-sm-12'><%= translate_helper('customer_eligibilities') %></h4>");
+        
+        var yes_label = "<%= translate_helper('yes') %>";
+        var no_label = "<%= translate_helper('no') %>";
+        for(var i=0; i<elig_data.length; i++) {
+          var elig = elig_data[i];
+          var new_elig_tags = '<div class="form-group col-sm-12">' +
+            '<label class="col-sm-3">' +
+              elig.description +
+            '</label>' +
+            '<div class="col-sm-9">' +
+              (elig.eligible ? yes_label : no_label) +
+            '</div>' +
+          '</div>';
+          $('.customer-eligibilities').append(new_elig_tags);
+        }
+      }
+    }
+
+    $('.trip_form #customer_name').bind('railsAutocomplete.select', function(event, data){ 
+      $('#view_customer').show();
+      popupFlagMessage(data.item.message);
+
+      if ($("#trip_group").val() == "true") {
+        $("li.passengers").hide();
+        $("li.group_size").show();
+      } else {
+        $("li.passengers").show();
+        $("li.group_size").hide();
+      }
+
+      var customerId = data.item.id;
+      $('.address-search input[name=customer_id]').val(customerId);
+
+      $.post("<%= data_for_trip_customers_path %>", {
+          customer_id: customerId
+        }, function(customerData) {
+          updateUponCustomerChange(customerData);
+        });
+    });
+
+    $('#view_customer').click(function(e) {
+      e.preventDefault();
+      var customerId = $('#repeating_trip_customer_id').val();
+      if(customerId && customerId.length > 0) {
+        var win = window.open("<%= customers_url %>" + '/' + customerId, '_blank');
+        win.focus();
+      }
+    });
+  });
+<% end %>
