+++ app/views/repeating_trips/_list.html.haml
@@ -1,0 +1,25 @@
+%table.table.table-striped.table-bordered
+  %thead
+    %tr
+      %th
+      %th= translate_helper("customer_column")
+      %th= translate_helper("pickup_address_column") 
+      %th= translate_helper("pickup_time_column") 
+      %th= translate_helper("dropoff_address_column")
+      %th= translate_helper("appointment_time_column") 
+  %tbody
+    - @trips.each do |trip|
+      %tr
+        %td
+          = link_to(translate_helper("view"), repeating_trip_path(trip))
+        %td
+          = link_to trip.customer.name, trip.customer, :title => translate_helper("view_customer_record_link_text") 
+        %td
+          = trip.pickup_address.try :address_text
+        %td
+          = format_time_for_listing trip.pickup_time
+        %td
+          = trip.dropoff_address.try :address_text
+        %td
+          = format_time_for_listing trip.appointment_time
+
