+++ app/views/repeating_trips/_repetition_panel.html.haml
@@ -3,7 +3,7 @@
     %h3.panel-title
       = translate_helper(:repetition)
   .panel-body{style: 'padding: 5px;'}
-    = render 'trips/repetition_fields', f: f
+    = render 'repeating_trips/repetition_fields', f: f
     
     .form-group.col-sm-12
       .col-sm-3
@@ -17,6 +17,6 @@
         = f.collection_select :repetition_driver_id, @drivers, :id, :name, { :include_blank => true }, { class: 'form-control' }
     .form-group.col-sm-12
       .col-sm-3
-        = f.label :repetition_customer_informed, translate_helper("mark_as_called_back")
+        = f.label :customer_informed, translate_helper("mark_as_called_back")
       .col-sm-9
-        = f.check_box :repetition_customer_informed
+        = f.check_box :customer_informed
