+++ app/views/repeating_trips/edit.html.haml
@@ -1,0 +1,4 @@
+= render 'trips/form', edit_mode: :edit 
+
+= render 'trips/javascript' 
+
