+++ app/views/repeating_trips/new.html.haml
@@ -1,0 +1,3 @@
+= render 'trips/form', edit_mode: :new
+
+= render 'trips/javascript'
