+++ app/views/repeating_trips/show.html.haml
@@ -1,0 +1,5 @@
+= render 'trips/form', edit_mode: :show 
+
+:javascript
+  $(".trip_form :input").attr("disabled","disabled");
+
