--- app/views/trips/_form.html.erb
@@ -1,37 +1,0 @@
-<%= render 'shared/highlight_active_tab_js', is_primary_nav: true, tab_class: 'trips' %>
-<% edit_mode = :edit if !defined? edit_mode%>
-<%= form_for @trip, :html => (@remote ? {"data-remote" => true, class: 'trip_form'} : {class: 'trip_form'} ) do |f| %>
-  <%= render 'form_header', edit_mode: edit_mode, f: f %>
-  <% if @trip.errors.any? %>
-    <div id="error_explanation">
-      <h2><%= translate_helper("trip_form_error_message", count: @trip.errors.count)%></h2>
-
-      <ul>
-      <% @trip.errors.full_messages.each do |msg| %>
-        <li><%= msg %></li>
-      <% end %>
-      </ul>
-    </div>
-  <% end %>
-
-  <!-- the customer id needs to be passed in this action -->
-  <%= f.hidden_field :customer_id %>
-  <%= f.hidden_field :direction %>
-  <%= hidden_field_tag "trip[outbound_trip_id]", @trip.outbound_trip.try(:id) %>
-
-  <div class="col-sm-12 col-md-6">
-    <%= render 'customer_panel', f: f, edit_mode: edit_mode %>
-    <%= render 'trip_parameters_panel', f: f, edit_mode: edit_mode %>
-    <%= render 'eligibility_accommodation_panel', f: f, edit_mode: edit_mode %>
-  </div>
-  <div class="col-sm-12 col-md-6">
-    <%= render 'repetition_panel', f: f %>
-    <%= render 'other_panel', f: f %>
-    <% if edit_mode != :new %>
-      <%= render 'assignment_panel', f: f %>
-    <% end %>
-  </div>
-  
-<% end %>
-
-<%= render 'shared/hide_invisible_form_fields_js', model_name: "trip", table_name: 'trips', provider_id:  current_provider.try(:id) %>
