+++ app/views/trips/_list.html.haml
@@ -14,4 +14,4 @@
     %th= translate_helper("trip_result_column") 
 
   - @trips.each do |trip|
-    = render "row", :trip => trip 
+    = render "trips/row", :trip => trip
