+++ app/views/trips/_trip_parameters_panel.html.haml
@@ -3,7 +3,7 @@
     %h3.panel-title
       = translate_helper(:trip_parameters)
   .panel-body{style: 'padding: 5px;'}
-    - if edit_mode == :show && @trip.is_linked?
+    - if !@is_repeating_trip && edit_mode == :show && @trip.is_linked?
       .form-group.col-sm-12
         .col-sm-3
           = label_tag :linked_trip, @trip.is_return? ?  translate_helper("outbound_trip_label") : translate_helper("return_trip_label")
@@ -32,11 +32,12 @@
       .col-sm-9
         = f.text_field :appointment_time, :value => @trip.appointment_time.try(:to_s, :form), class: 'datetime-picker form-control', readonly: true
 
-    .form-group.col-sm-12
-      .col-sm-3
-        = f.label :direction, translate_helper("trip_direction")
-      .col-sm-9
-        = @trip.is_return? ? translate_helper("return_trip") : translate_helper("outbound_trip")
+    - unless @is_repeating_trip
+      .form-group.col-sm-12
+        .col-sm-3
+          = f.label :direction, translate_helper("trip_direction")
+        .col-sm-9
+          = @trip.is_return? ? translate_helper("return_trip") : translate_helper("outbound_trip")
 
     .form-group.col-sm-12.place-picker-div
       .col-sm-3
@@ -122,7 +123,7 @@
               = f.text_field :attendant_count, class: 'form-control'
               &nbsp;
               = translate_helper("attendants")
-    - if edit_mode == :show
+    - if !@is_repeating_trip && edit_mode == :show
       .form-group.col-sm-12
         .col-sm-3
           = f.label :drive_distance, translate_helper("drive_distance")
