--- app/views/trips/edit.html.erb
@@ -1,4 +1,0 @@
-<%= render 'form', edit_mode: :edit %>
-
-<%= render 'javascript' %>
-
