+++ app/views/trips/edit.html.haml
@@ -1,0 +1,4 @@
+= render 'form', edit_mode: :edit 
+
+= render 'javascript' 
+
