--- app/views/trips/new.html.erb
@@ -1,3 +1,0 @@
-<%= render 'form', edit_mode: :new %>
-
-<%= render 'javascript' %>
