+++ app/views/trips/new.html.haml
@@ -1,0 +1,3 @@
+= render 'form', edit_mode: :new
+
+= render 'javascript'
