+++ config/routes.rb
@@ -71,6 +71,8 @@ Rails.application.routes.draw do
       end
     end
 
+    resources :repeating_trips 
+
     resources :providers, :except => [:edit, :update, :destroy] do
       post :change_role
       post :delete_role
